release: 0.64.0
This update is big. Users will find that I changed the entire API except for traits and algorithms. Analyzing the digraph `op` and `gen` implementations for the standard library collections, I settled on two main digraph representations:

- `adjacency_list`
- `adjacency_list_weighted`

These representations offer a good balance between performance and ergonomics. All applicable traits are implemented and tested for these representations. The `adjacency_matrix` representation is still available for dense unweighted digraphs, but blocks are now stored in a `Vec` instead of an array.

The next minor release will add a user guide to the documentation.